requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
dashscope>=1.14.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0